Complete utils modules refactoring
